Refactor elf_specs.h to use the original glibc 2.22 elf.h.
Previously, elf_specs.h was an old(!) copy of glibc's elf.h with a few
additions accumulated over time.

This commit introduces a clear separation between an unmodified (save
one explanatory comment and a removed #include) copy of elf.h called
glibc_elf.h (elf.h already exists here), and elf_specs.h, which includes
the former file and only contains the required changes.

Updating glibc_elf.h in the future should be as easy as copying the file
from glibc and potentially deleting obsoleted parts of elf_specs.h.
